+++ protocol/whist/logging/logging.c
@@ -131,11 +131,11 @@ static WhistMutex crash_handler_mutex;
  */
 static bool crash_handler_mutex_active;
 
-static void (*external_logger_callback)(unsigned int level, const char* line, int session_id) = NULL;
+static void (*external_logger_callback)(unsigned int level, const char* line, long session_id) = NULL;
 
-static int session_id;
+static long session_id;
 
-void whist_set_session_id (int new_session_id) {
+void whist_set_session_id (long new_session_id) {
     session_id = new_session_id;
 }
